Only use hwclock when required.
Now we check for PLATFORM and we break out of the script if we are not QASMP.
And the script is only used to save the system clock into the RTC at
shutdown or reboot time;
during initialization the kernel loads RTC into system clock.
Keeps our bring up fast.