Changes for building in build tree.
-I's and -L's are pulled from source tree and/or build tree where possible.
This commit is partial (which is perfectly fine.)
I am staging commits as some depend upon other packages.
This is the simplest group.
Note the use of -I$(somePackage_top_srcdir) which is prevalent and
we define these xxx_top_srcdir variables in the xxx.mk file.
DESTDIR is used some places without being defined and that is OK,
it is like it is not being used (yet.)
Some places we use a macro rather than  INSTALL_DIR in anticipation of a
slightly more flexible approach.